rest-api/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

langchain
langgraph
openai
sqlalchemy
mysql-connector-python

pandas
numpy